+++ src/Splunk.OpenTelemetry.AutoInstrumentation/ConfigurationKeys.cs
@@ -96,6 +96,11 @@ internal static class ConfigurationKeys
             public const string Enabled = "SPLUNK_SNAPSHOT_PROFILER_ENABLED";
 
             /// <summary>
+            /// Configuration key for enabling high resolution timer on Windows.
+            /// </summary>
+            public const string HighResolutionTimerEnabled = "SPLUNK_SNAPSHOT_HIGH_RES_TIMER_ENABLED";
+
+            /// <summary>
             /// Configuration key for snapshots sampling interval in ms.
             /// </summary>
             public const string SamplingIntervalMs = "SPLUNK_SNAPSHOT_SAMPLING_INTERVAL";