+++ tools/MatrixHelper/Instrumentation.cs
@@ -21,17 +21,17 @@ namespace MatrixHelper;
 internal class Instrumentation
 {
     public Instrumentation(string key, InstrumentedComponent instrumentedComponent, string? description, string stability, string support, SignalsList[] signalsList)
-    : this(new[] { key }, new[] { instrumentedComponent }, description, stability, support, Array.Empty<Dependency>(), signalsList, Array.Empty<Setting>())
+    : this([key], [instrumentedComponent], description, stability, support, [], signalsList, [])
     {
     }
 
     public Instrumentation(string key, InstrumentedComponent instrumentedComponent, string? description, string stability, string support, SignalsList signalsList)
-        : this(new[] { key }, new[] { instrumentedComponent }, description, stability, support, Array.Empty<Dependency>(), new[] { signalsList }, Array.Empty<Setting>())
+        : this([key], [instrumentedComponent], description, stability, support, [], [signalsList], [])
     {
     }
 
     public Instrumentation(string key, InstrumentedComponent instrumentedComponent, string? description, string stability, string support, Dependency dependency, SignalsList[] signalsList)
-        : this(new[] { key }, new[] { instrumentedComponent }, description, stability, support, new[] { dependency }, signalsList, Array.Empty<Setting>())
+        : this([key], [instrumentedComponent], description, stability, support, [dependency], signalsList, [])
     {
     }