+++ tools/MatrixHelper/MetadataData.cs
@@ -25,11 +25,11 @@ internal static class MetadataData
         return new AllInOne(
             component: "Splunk Distribution of OpenTelemetry .NET",
             version: typeof(MetadataData).Assembly.GetCustomAttribute<AssemblyInformationalVersionAttribute>()!.InformationalVersion.Split('+')[0],
-            dependencies: new Dependency[]
-            {
-                new("OpenTelemetry .NET", "https://github.com/open-telemetry/opentelemetry-dotnet", null, "1.12.0", "stable"),
-                new("OpenTelemetry .NET Automatic Instrumentation", "https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation", null, "1.13.0-beta.1", "beta"),
-            },
+            dependencies:
+            [
+                new("OpenTelemetry .NET", "https://github.com/open-telemetry/opentelemetry-dotnet", null, "1.14.0", "stable"),
+                new("OpenTelemetry .NET Automatic Instrumentation", "https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation", null, "1.13.0", "stable")
+            ],
             settings: SettingsData.GetSettings(),
             instrumentations: InstrumentationData.GetInstrumentations(),
             resourceDetectors: ResourceDetectorsData.GetResourceDetectors());