+++ tools/MatrixHelper/TracesList.cs
@@ -18,5 +18,5 @@ namespace MatrixHelper;
 
 internal class TracesList : SignalsList
 {
-    public Trace[] Traces { get; set; } = Array.Empty<Trace>();
+    public Trace[] Traces { get; set; } = [];
 }